layout: fab toggle moved to bottom bar
